moved TaskDTO from logic to api
